\name{c_list}
\alias{c_list}
\title{cluster_method_list}
\usage{
c_list(dat, 30)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
This function uses the Network Modularity Optimazation Cluster Method to cluster the cells. It returns 
the list of indices and name of cells after clustering. 
}
