\name{GCluster}
\alias{GCluster}

\title{
 graphical based cluster
}

\description{
 graphical based cluster
}

\usage{
  GCluster(dat=dat, wt=4, k=NULL, method="louvain");
}

\arguments{
   \item{dat}{data.frame or matrix}
   \item{wt}{weight}
   \item{k}{cluster number}
   \item{method}{The method should be one of louvain, spinglass, 
   fast_greedy, infomap or label_prop.}
}


\examples{

}

