\name{m2p}
\alias{m2p}

\title{
 quick calculate corralation coefficent and p value with big matrix
}

\description{
 quick calculate corralation coefficent and p value with big matrix
}

\usage{
  m2p(m, alternative="two-sided");
}

\arguments{
   \item{m}{matrix, row is sample and column is gene}
   \item{alternative}{two-sided, less or great}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

