\name{scale.v}
\alias{scale.v}

\title{
 scale function
}

\description{
 scale function
}

\usage{
  scale.v(x, start.v, end.v)
}

\arguments{
   \item{x}{numerical vector}
   \item{start.v}{minimum of scale}
   \item{end.v}{maximum of scale}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

