\name{tjGCluster}
\alias{tjGCluster}

\title{
 trajectory analysis
}

\description{
 trajectory analysis
}

\usage{
  tjGCluster(m, from=NULL, to=NULL);
}

\arguments{
   \item{m}{data.frame or matrix or igraph object}
   \item{from}{start vetex}
   \item{to}{end vetex}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

