\name{GCluster}
\alias{GCluster}

\title{
 graph based cluster 2
}

\description{
 graph based cluster 2
}

\usage{
  GCluster(dat=dat, wt=4, k=NULL, method="louvain");
}

\arguments{
   \item{dat}{data.frame or matrix}
   \item{wt}{weight}
   \item{k}{cluster number}
   \item{method}{The method should be one of louvain, spinglass, 
   fast_greedy, infomap or label_prop.}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

