\name{m_list}
\alias{m_list}
\title{window_method_list}
\usage{
m_list(coordinates, l)
}
\arguments{
\item{coordinates}{Array of coordinates of the data}
\item{l}{Number of windows}
}
\value{
A list of indices and name of cells.
}
\description{
This function uses the window-based method to merge the cells. It returns 
the list of indices and name of cells after clustering. 
}
