\name{mgGCluster}
\alias{mgGCluster}

\title{
 merge cluster
}

\description{
 merge cluster
}

\usage{
  mgGCluster(x, l, rename=TRUE);
}

\arguments{
   \item{x}{vactor as cluster ID}
   \item{l}{list, each object is vector. all clusters will be renamed as the first one}
   \item{rename}{if TRUE, the ID will be renamed.}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

